pyproject.toml
[project]
name = "stepic-lc"
version = "0.1.0"
description = "LangChain Learning Project - Console Chat Bot and Prompt Engineering Examples"
authors = [
    {name = "Developer", email = "developer@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "langchain-community>=0.3.31",
    "langchain-huggingface>=0.3.1",
    "sentence-transformers>=5.1.2",
    "chromadb>=1.3.5",
    "pypdf>=6.4.2",
    "pymupdf>=1.26.5",
    "beautifulsoup4>=4.14.3",
    "selenium>=4.36.0",
    "unstructured>=0.18.3",
    "markdown>=3.9",
    "docx2txt>=0.9",
    "jq>=1.10.0",
    "faiss-cpu>=1.13.0",
    "qdrant-client>=1.16.1",
    "langchain-qdrant>=0.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
